chore: update branding of river.com/delegate to Towns
