+++ src/components/claim/connect-wallet.tsx
@@ -1,8 +1,8 @@
 import { cn } from '@/lib/utils'
 import { useWeb3Modal } from '@web3modal/wagmi/react'
+import Image from 'next/image'
 import { Button } from '../ui/button'
 import { Typography } from '../ui/typography'
-import BeaverClaimAscii from './beaver-claim-ascii'
 
 export const ConnectWallet = () => {
   const { open } = useWeb3Modal()
@@ -15,9 +15,7 @@ export const ConnectWallet = () => {
         'min-h-screen items-center',
       )}
     >
-      <div className="h-[310px]">
-        <BeaverClaimAscii className="size-full"  />
-      </div>
+      <Image src="/images/towns-t-logo.png" alt="TOWNS T Logo" width={269} height={269} />
       <div className="flex flex-col items-center gap-2">
         <Typography
           as="h1"
