+++ src/components/towns-logo.tsx
@@ -1,0 +1,13 @@
+
+export const TownsLogo = (props: React.SVGProps<SVGSVGElement>) => {
+  return (
+    <svg width="100" height="26" viewBox="0 0 100 26" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M23.2447 2.18129C22.6688 1.51385 21.8168 1.09138 20.8661 1.09138H8.11555C7.86497 1.09138 7.62124 1.12073 7.38762 1.17618C6.69112 1.24726 6.03963 1.59187 5.58768 2.15143C5.58768 2.15143 0.97317 8.22442 0.590854 8.69777C0.208538 9.17111 0 9.76119 0 10.3696V14.2783C0 14.984 0.280334 15.6607 0.77933 16.1597L1.41225 16.7927C1.87095 17.2514 2.49307 17.509 3.14177 17.509C4.49261 17.509 5.58768 18.6041 5.58768 19.955V21.9528C5.58768 22.6585 5.68635 23.5611 6.30804 24.2825C6.90942 24.9803 7.64672 25.3344 8.41934 25.3344H89.5839C91.1713 25.3344 92.8397 24.463 93.7192 23.8448C94.5987 23.2266 97.5903 19.2929 98.4733 18.0598C99.3562 16.8267 100 15.4393 100 13.8318V12.6185C100 11.5765 99.6777 10.6075 99.127 9.80636C99.6891 9.22055 100.027 8.4082 99.9552 7.47704C99.7553 4.87504 98.8588 2.98815 97.1817 1.86984C95.6753 0.865291 93.8819 0.756391 92.7204 0.690289L92.6614 0.686936H85.9185C84.1485 0.686936 82.4959 1.39788 81.2574 2.5946C80.7055 1.69299 79.7115 1.09138 78.577 1.09138H74.7455C73.6532 1.09138 72.6913 1.64897 72.1287 2.49502C71.5468 1.6184 70.5645 1.09138 69.5123 1.09138H64.5919C63.9916 1.09138 63.4308 1.25977 62.9539 1.55189C62.4821 1.26162 61.923 1.09138 61.3122 1.09138H58.319C56.9771 1.09138 55.7835 1.94403 55.3484 3.21341L55.3461 3.2199C54.9175 1.95804 53.7281 1.09138 52.3732 1.09138H50.2491C48.8942 1.09138 47.7048 1.95804 47.2761 3.21991L47.2739 3.21341C46.8388 1.94403 45.6452 1.09138 44.3033 1.09138H41.31C40.4074 1.09138 39.6178 1.46318 39.0557 2.04472L38.9727 2.03386C37.8503 1.17502 36.4488 0.665649 34.9356 0.665649H27.4853C25.9034 0.665649 24.4251 1.22052 23.2447 2.18129ZM34.9358 2.74215H27.4855C26.2509 2.74215 25.1015 3.23173 24.2287 4.08319C23.3773 4.93465 22.8877 6.1054 22.909 7.31872V13.8324C22.8877 15.0457 23.3773 16.1951 24.2287 17.0679C25.1015 17.9193 26.2722 18.4089 27.4855 18.4089H34.9358C37.4476 18.4089 39.5124 16.3442 39.5124 13.8324V7.31872C39.5124 4.78564 37.4476 2.74215 34.9358 2.74215ZM33.5309 12.172C33.5309 13.0022 32.8497 13.6834 32.0195 13.6834H30.3805C29.5503 13.6834 28.8692 13.0022 28.8692 12.172V8.97906C28.8692 8.14889 29.5503 7.46773 30.3805 7.46773H32.0195C32.8497 7.46773 33.5309 8.14889 33.5309 8.97906V12.172ZM8.11575 3.16787C7.52794 3.16787 7.05143 3.64439 7.05143 4.23219V7.21229C7.05143 7.8001 7.52794 8.27661 8.11575 8.27661H10.4291C11.0169 8.27661 11.4934 8.75312 11.4934 9.34093V17.1553C11.4934 17.7431 11.97 18.2196 12.5578 18.2196H16.4243C17.0121 18.2196 17.4886 17.7431 17.4886 17.1553V9.34093C17.4886 8.75312 17.9651 8.27661 18.5529 8.27661H20.8663C21.4541 8.27661 21.9306 7.8001 21.9306 7.21229V4.23219C21.9306 3.64439 21.4541 3.16787 20.8663 3.16787H8.11575ZM57.3124 3.88708C57.4598 3.45686 57.8644 3.16787 58.3192 3.16787H61.3125C62.0468 3.16787 62.5605 3.89389 62.3161 4.58634L57.8095 17.3582C57.6593 17.7838 57.2571 18.0684 56.8058 18.0684H53.0633C52.8255 18.0684 52.603 17.9774 52.4348 17.8237C52.2132 17.6212 52.17 17.301 52.1298 17.0032L52.1292 16.9987C52.0005 16.0451 51.8449 15.5155 51.3336 15.5155C50.8203 15.5155 50.6276 16.0389 50.4899 17.0056L50.4865 17.0292C50.4458 17.3158 50.4025 17.6207 50.1919 17.8178C50.0236 17.9752 49.7985 18.0684 49.5574 18.0684H45.8169C45.3656 18.0684 44.9634 17.7838 44.8132 17.3582L40.3066 4.58635C40.0622 3.89389 40.5759 3.16787 41.3102 3.16787H44.3035C44.7583 3.16787 45.1628 3.45686 45.3103 3.88708L46.4419 7.18829C46.7882 8.19848 48.2403 8.12633 48.4848 7.08679L49.2132 3.98859C49.3263 3.50774 49.7553 3.16787 50.2493 3.16787H52.3734C52.8673 3.16787 53.2964 3.50774 53.4094 3.98859L54.1379 7.08678C54.3823 8.12632 55.8345 8.19848 56.1808 7.18829L57.3124 3.88708ZM74.7457 3.16787C74.1579 3.16787 73.6814 3.64439 73.6814 4.23219V5.06338C73.6814 6.11793 72.3124 6.53037 71.7299 5.65135L70.3996 3.64423C70.2025 3.34674 69.8694 3.16787 69.5125 3.16787H64.5921C64.0042 3.16787 63.5277 3.64439 63.5277 4.23219V17.0253C63.5277 17.6131 64.0042 18.0896 64.5921 18.0896H68.5939C69.1817 18.0896 69.6582 17.6131 69.6582 17.0253V14.8321C69.6582 13.7382 71.1104 13.3559 71.6489 14.3081L73.4822 17.5493C73.671 17.8832 74.025 18.0896 74.4086 18.0896H78.5772C79.165 18.0896 79.6415 17.6131 79.6415 17.0253V4.23219C79.6415 3.64439 79.165 3.16787 78.5772 3.16787H74.7457ZM87.7572 9.17895C87.7939 9.18955 87.8314 9.19318 87.8695 9.1958C88.1957 9.21821 89.8997 9.33485 91.964 9.46865C92.7089 9.52945 93.508 9.59025 94.3071 9.65105L94.3073 9.65107L94.3075 9.65109C94.6273 9.67542 94.9471 9.69975 95.2634 9.72408C96.7535 9.8518 97.9242 11.1077 97.9242 12.619V13.8324C97.9242 16.3442 95.8807 18.3876 93.3476 18.3876H85.5568C83.045 18.3876 81.3634 16.3442 81.3634 13.8324V13.5131C81.3634 12.9253 81.8399 12.4487 82.4277 12.4487H86.3124C86.6286 12.4487 87 12.5871 87.3715 12.7255C87.7429 12.8638 88.1143 13.0022 88.4305 13.0022H91.2403C91.8599 13.0022 92.0387 12.1204 91.5042 11.9708C91.4707 11.9614 91.4362 11.9578 91.4014 11.9551C91.0835 11.9298 89.3406 11.7896 87.3023 11.5973C87.087 11.5766 86.8684 11.5559 86.6481 11.535C85.7364 11.4486 84.7963 11.3596 83.9391 11.2567C82.4703 11.0864 81.3421 9.8518 81.3634 8.38304V7.34001C81.3421 6.12668 81.8317 4.95593 82.6832 4.10448C83.5559 3.25302 84.7054 2.76343 85.9187 2.76343H92.6026C95.0378 2.90202 97.5467 3.22475 97.8855 7.63656C97.9306 8.22264 97.4477 8.70234 96.8599 8.70234H92.9721C92.6531 8.70234 92.2785 8.5693 91.904 8.43626C91.5294 8.30322 91.1549 8.17018 90.8359 8.17018H88.026C87.4083 8.17018 87.2288 9.0266 87.7572 9.17895Z"
+        fill="white"
+      />
+    </svg>
+  )
+}
